requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-resource-manager==1.14.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-APScheduler==1.13.0
cryptography==41.0.7
openai==1.97.0
beautifulsoup4==4.12.2
playwright==1.40.0
psutil==5.9.6
psycopg[binary]==3.2.9 